+++ src/query-change-detector.js
@@ -144,7 +144,7 @@ class QueryChangeDetector {
             }
         } else {
             // U1 doc not matched and also not matches now
-            if (!options.skip && !options.limit && !wasDocInResults && !doesMatchNow) {
+            if (!options.skip && !wasDocInResults && !doesMatchNow) {
                 DEBUG && _debugMessage(this, 'U1', docData);
                 return false;
             }
@@ -184,6 +184,8 @@ class QueryChangeDetector {
         }
 
         // if no optimisation-algo matches, return mustReExec:true
+        DEBUG && console.log(this, 'NO_MATCH', docData);
+
         return true;
     }
 }