+++ test/unit/query-change-detector.test.js
@@ -415,9 +415,9 @@ config.parallel('query-change-detector.test.js', () => {
                 // it should find the same order with pouchdb
                 const pouchResult = await col.pouch.find(
                     col
-                    .find()
-                    .where('passportId')
-                    .ne('foobar3').toJSON()
+                        .find()
+                        .where('passportId')
+                        .ne('foobar3').toJSON()
                 );
                 assert.deepEqual(
                     docs.map(d => d.id),
@@ -462,9 +462,9 @@ config.parallel('query-change-detector.test.js', () => {
                 // it should find the same order with pouchdb
                 const pouchResult2 = await col.pouch.find(
                     col
-                    .find()
-                    .where('passportId')
-                    .ne('foobar3').toJSON()
+                        .find()
+                        .where('passportId')
+                        .ne('foobar3').toJSON()
                 );
                 assert.deepEqual(
                     lastResult,
@@ -522,7 +522,90 @@ config.parallel('query-change-detector.test.js', () => {
             col.database.destroy();
         });
     });
-    describe('e', () => {
-        //    it('e', () => process.exit());
+    it('BUG: no optimisation for irrelevant insert', async () => {
+        const schema = {
+            title: 'messages schema',
+            description: 'describes a message',
+            version: 0,
+            keyCompression: false,
+            type: 'object',
+            properties: {
+                id: {
+                    type: 'string',
+                    primary: true
+                },
+                text: {
+                    type: 'string'
+                },
+                time: {
+                    type: 'number',
+                    index: true
+                },
+                read: {
+                    description: 'true if was read by the reciever',
+                    type: 'boolean'
+                },
+                sender: {
+                    type: 'string',
+                    ref: 'users'
+                },
+                reciever: {
+                    type: 'string',
+                    ref: 'users'
+                }
+            },
+            compoundIndexes: [
+                ['sender', 'time'],
+                ['reciever', 'time']
+            ],
+            required: [
+                'text',
+                'time',
+                'read',
+                'sender',
+                'reciever'
+            ]
+        };
+        const col = await humansCollection.createBySchema(schema);
+        const user1 = '1';
+        const user2 = '2';
+        const getQuery = () => col.findOne({
+            $or: [
+                {
+                    sender: {
+                        $eq: user1
+                    },
+                    reciever: {
+                        $eq: user2
+                    }
+                },
+                {
+                    sender: {
+                        $eq: user2
+                    },
+                    reciever: {
+                        $eq: user1
+                    }
+                }
+            ]
+        }).sort('-time');
+
+        await getQuery().exec();
+        const countBefore = getQuery()._execOverDatabaseCount;
+
+        // insert something that does not match
+        await col.insert({
+            id: AsyncTestUtil.randomString(10),
+            text: AsyncTestUtil.randomString(10),
+            time: AsyncTestUtil.randomNumber(1, 1000),
+            read: false,
+            sender: '3',
+            reciever: '4'
+        });
+
+        await getQuery().exec();
+        const countAfter = getQuery()._execOverDatabaseCount;
+        assert.equal(countBefore, countAfter);
+        col.database.destroy();
     });
 });