+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Bugfixes:
   - custom server path not working on `server`-plugin [#1447](https://github.com/pubkey/rxdb/issues/1447) Thanks [@dstudzinski](https://github.com/dstudzinski)
   - Fix CORS headers when the request's credentials mode is 'include' [#1450](https://github.com/pubkey/rxdb/issues/1450) Thanks [@dstudzinski](https://github.com/dstudzinski)
 
+Other:
+  - Improved `QueryChangeDetection` to not run on irrelevant changes
+
 ### 8.4.0 (1 September 2019)
 
 Bugfixes: